Adds PyTorch model management (#394)
Adding PyTorch model support through two main mechanisms:
(1) Generic PyTorch model management in Elasticsearch
(2) Interoperability with HuggingFace transformers and model hub

This change aims to provide the foundations of this support and all of the basic functionality that one would need to get up and running. Note that this functionality is available in Elasticsearch 8.x builds only.