Improves the options and instructions for install
Eland's dependencies (and transitive dependencies) require quite a few
prerequisites like numpy, libzip, libjpeg, etc. On a normal macOS
machine this isn't so bad and things tend to "just work", however on
vanilla Debian machines such as the default Debian 10 image on GCP,
there are quite a few prerequisites to work through.

This change introduces instructions that explicitly help with Debian
installations, and a second option to build a Debian Docker container in
order to avoid having to figure out the dependencies needed for your
particular distribution.

The Docker container version is really only useful when you want to just
run the installed scripts but it provides a very easy mechanism if
people just want to, for example, upload PyTorch NLP models and don't
actually want to use any of the other eland functionality.